requirements.txt
Django==4.2.4
django-environ==0.10.0
djangorestframework==3.14.0
django-cors-headers==4.2.0
gunicorn==21.2.0
